+++ vCardLib.Tests/DeserializerTests/V4DeserializerTests.cs
@@ -1,6 +1,7 @@
 ﻿using System;
 using NUnit.Framework;
 using vCardLib.Deserializers;
+using vCardLib.Models;
 
 namespace vCardLib.Tests.DeserializerTests
 {
