+++ vCardLib.Tests/SerializerTests/V2SerializerTests.cs
@@ -1,5 +1,6 @@
 using System;
 using NUnit.Framework;
+using vCardLib.Models;
 using vCardLib.Serializers;
 
 namespace vCardLib.Tests.SerializerTests
