+++ vCardLib/Deserializers/V4Deserializer.cs
@@ -1,4 +1,5 @@
 ﻿using System;
+using vCardLib.Models;
 
 namespace vCardLib.Deserializers
 {
