+++ vCardLib/Enums/EmailType.cs
@@ -1,0 +1,16 @@
+﻿namespace vCardLib.Enums
+{
+    /// <summary>
+    /// Various email address types in a vCard
+    /// </summary>
+    public enum EmailType
+    {
+        Work,
+        Internet,
+        Home,
+        AOL,
+        Applelink,
+        IBMMail,
+        None
+    }
+}
