+++ vCardLib/Enums/PhotoEncoding.cs
@@ -1,0 +1,11 @@
+﻿namespace vCardLib.Enums
+{
+    /// <summary>
+    /// Supported ecnoding types
+    /// </summary>
+    public enum PhotoEncoding
+    {
+        JPEG,
+        GIF
+    }
+}
