+++ vCardLib/Enums/PhotoType.cs
@@ -1,0 +1,11 @@
+﻿namespace vCardLib.Enums
+{
+    /// <summary>
+    /// Support image format types
+    /// </summary>
+    public enum PhotoType
+    {
+        Image,
+        URL
+    }
+}
