+++ vCardLib/Models/EmailAddress.cs
@@ -1,4 +1,4 @@
-﻿using System.Net.Mail;
+﻿using vCardLib.Enums;
 
 namespace vCardLib.Models
 {
@@ -10,24 +10,11 @@ namespace vCardLib.Models
         /// <summary>
         /// The email address
         /// </summary>
-        public MailAddress Email { get; set; }
+        public string Email { get; set; }
+        
         /// <summary>
         /// The email address type
         /// </summary>
         public EmailType Type { get; set; }
-    }
-
-    /// <summary>
-    /// Various email address types in a vCard
-    /// </summary>
-    public enum EmailType
-    {
-        Work,
-        Internet,
-        Home,
-        AOL,
-        Applelink,
-        IBMMail,
-        None
     }
 }
