+++ vCardLib/vCardLib.csproj
@@ -1,6 +1,5 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
     <PropertyGroup>
-        <TargetFrameworks>netstandard2.0</TargetFrameworks>
         <Version>2.2.6</Version>
         <Copyright>Copyright 2019</Copyright>
         <PackageReleaseNotes>Fix bugs</PackageReleaseNotes>
@@ -17,8 +16,6 @@
         <RepositoryType>git</RepositoryType>
         <PackageTags>vcard, vcardlib</PackageTags>
         <PackageVersion>3.0.0</PackageVersion>
+        <TargetFramework>netstandard1.3</TargetFramework>
     </PropertyGroup>
-    <ItemGroup>
-        <PackageReference Include="CoreCompat.System.Drawing.v2" Version="5.2.0-preview1-r131" />
-    </ItemGroup>
 </Project>
